arch: move paths.rs from app to infra layer (#187)
Path resolution (state_dir, log_dir, bus socket, reminders) is an
infrastructure concern, not application logic. Moving it to infra
removes the config → app layer violation flagged by archlint.

config.rs re-exports from infra::paths instead of app::paths —
all call sites use config:: and are unchanged.
