+++ src/infra/mod.rs
@@ -3,4 +3,5 @@
 pub mod dto;
 pub mod memory_bus;
 pub mod memory_store;
+pub mod paths;
 pub mod unix_bus;
